pyproject.toml
[project]
name = "bevfusionv2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "matplotlib==3.9.2",
    "numpy==1.26.4",
    "nuscenes-devkit==1.2.0",
    "opencv-python==4.10.0.84",
    "pandas==2.2.3",
    "pyyaml==6.0.2",
    "torch==2.9.0",
    "torchaudio==2.9.0",
    "torchvision==0.24.0",
    "tqdm==4.66.5",
]
